Moved functions from Generator to utils folder
